More detailed results 6

+++ CppScripts/BBBhw/GPIO.h
@@ -64,7 +64,7 @@ private:// Variables
 	static void *pru_int;       // Points to start of PRU memory.
 	//static int chunk;
 	static unsigned int *sharedMem_int,*pru0dataMem_int,*pru1dataMem_int;
-	unsigned int valCarryOnCycleCountPRU=0; // 32 bits
+	unsigned long long int valCarryOnCycleCountPRU=0; // 64 bits
 	// PRU timetagger
 	int retInterruptsPRU0;
 	int WaitTimeInterruptPRU0=2000000; // In microseconds
